fix worker import
